+++ agent/client_overview.php
@@ -372,13 +372,13 @@ $sql_asset_retired = mysqli_query(
                         $credential_uri_display = "<a href='$credential_uri'>" . truncate($credential_uri,40) . "</a><button class='btn btn-sm clipboardjs' type='button' title='$credential_uri' data-clipboard-text='$credential_uri'><i class='far fa-copy text-secondary'></i></button>";
                     }
                     $credential_uri_2 = sanitize_url($row['credential_uri_2']);
-                    $credential_username = nullable_htmlentities(decryptCredentialEntry($row['credential_username']));
+                    $credential_username = nullable_htmlentities(decryptCredentialEntry($row['credential_username'], $row['credential_client_id']));
                     if (empty($credential_username)) {
                         $credential_username_display = "-";
                     } else {
                         $credential_username_display = "$credential_username<button class='btn btn-sm clipboardjs' type='button' data-clipboard-text='$credential_username'><i class='far fa-copy text-secondary'></i></button>";
                     }
-                    $credential_password = nullable_htmlentities(decryptCredentialEntry($row['credential_password']));
+                    $credential_password = nullable_htmlentities(decryptCredentialEntry($row['credential_password'], $row['credential_client_id']));
                     $credential_otp_secret = nullable_htmlentities($row['credential_otp_secret']);
                     $credential_id_with_secret = '"' . $row['credential_id'] . '","' . $row['credential_otp_secret'] . '"';
                     if (empty($credential_otp_secret)) {