+++ agent/global_search.php
@@ -611,8 +611,8 @@ if (isset($_GET['query'])) {
                                 $credential_name = nullable_htmlentities($row['credential_name']);
                                 $credential_description = nullable_htmlentities($row['credential_description']);
                                 $credential_client_id = intval($row['credential_client_id']);
-                                $credential_username = nullable_htmlentities(decryptCredentialEntry($row['credential_username']));
-                                $credential_password = nullable_htmlentities(decryptCredentialEntry($row['credential_password']));
+                                $credential_username = nullable_htmlentities(decryptCredentialEntry($row['credential_username'], $row['credential_client_id']));
+                                $credential_password = nullable_htmlentities(decryptCredentialEntry($row['credential_password'], $row['credential_client_id']));
                                 $client_id = intval($row['client_id']);
                                 $client_name = nullable_htmlentities($row['client_name']);