+++ agent/modals/asset/asset_details.php
@@ -521,13 +521,13 @@ ob_start();
                         } else {
                             $credential_uri_display = "$credential_uri";
                         }
-                        $credential_username = nullable_htmlentities(decryptCredentialEntry($row['credential_username']));
+                        $credential_username = nullable_htmlentities(decryptCredentialEntry($row['credential_username'], $row['credential_client_id']));
                         if (empty($credential_username)) {
                             $credential_username_display = "-";
                         } else {
                             $credential_username_display = "$credential_username <button type='button' class='btn btn-sm clipboardjs' data-clipboard-text='$credential_username'><i class='far fa-copy text-secondary'></i></button>";
                         }
-                        $credential_password = nullable_htmlentities(decryptCredentialEntry($row['credential_password']));
+                        $credential_password = nullable_htmlentities(decryptCredentialEntry($row['credential_password'], $row['credential_client_id']));
                         $credential_otp_secret = nullable_htmlentities($row['credential_otp_secret']);
                         $credential_id_with_secret = '"' . $row['credential_id'] . '","' . $row['credential_otp_secret'] . '"';
                         if (empty($credential_otp_secret)) {