+++ agent/post/asset.php
@@ -54,8 +54,8 @@ if (isset($_POST['add_asset'])) {
     mysqli_query($mysqli,"INSERT INTO asset_interfaces SET interface_name = '01', interface_mac = '$mac', interface_ip = '$ip', interface_nat_ip = '$nat_ip', interface_ipv6 = '$ipv6', interface_primary = 1, interface_network_id = $network, interface_asset_id = $asset_id");
 
     if (!empty($_POST['username'])) {
-        $username = trim(mysqli_real_escape_string($mysqli, encryptCredentialEntry($_POST['username'])));
-        $password = trim(mysqli_real_escape_string($mysqli, encryptCredentialEntry($_POST['password'])));
+        $username = trim(mysqli_real_escape_string($mysqli, encryptCredentialEntry($_POST['username'], $client_id)));
+        $password = trim(mysqli_real_escape_string($mysqli, encryptCredentialEntry($_POST['password'], $client_id)));
 
         mysqli_query($mysqli,"INSERT INTO credentials SET credential_name = '$name', credential_username = '$username', credential_password = '$password', credential_asset_id = $asset_id, credential_client_id = $client_id");